+++ src/evm/protocol/uniswap_v2/mod.rs
@@ -1,5 +1,4 @@
 //! Uniswap V2 Decentralized Exchange
-pub mod events;
 pub mod reserve_price;
 pub mod state;
 pub mod tycho_decoder;