+++ src/protocol/state.rs
@@ -9,7 +9,6 @@
 //!  - `spot_price`: Returns the current spot price between two tokens.
 //!  - `get_amount_out`: Returns the amount of output tokens given an amount of input tokens.
 //!  - `delta_transition`: Applies a state delta to the protocol sim.
-//!  - `event_transition`: Applies an event transition to the protocol sim.
 //!  - `clone_box`: Clones the protocol sim as a trait object.
 //!  - `as_any`: Allows downcasting of the trait object.
 //!  - `eq`: Compares two protocol sims for equality.
@@ -51,7 +50,6 @@ use crate::{
     models::ERC20Token,
     protocol::{
         errors::{SimulationError, TransitionError},
-        events::{EVMLogMeta, LogIndex},
         models::GetAmountOutResult,
     },
 };
@@ -120,25 +118,6 @@ pub trait ProtocolSim: std::fmt::Debug + Send + Sync + 'static {
         tokens: Vec<ERC20Token>,
     ) -> Result<(), TransitionError<String>>;
 
-    /// Applies an event transition to the protocol's state.
-    ///
-    /// This method processes a protocol-specific event and modifies the protocol's state
-    /// accordingly.
-    ///
-    /// # Arguments
-    ///
-    /// * `protocol_event` - The event to apply to the protocol's state.
-    /// * `log` - Metadata about the EVM log that triggered the event.
-    ///
-    /// # Returns
-    ///
-    /// A `Result<(), TransitionError<LogIndex>>` indicating success or failure.
-    fn event_transition(
-        &mut self,
-        protocol_event: Box<dyn ProtocolEvent>,
-        log: &EVMLogMeta,
-    ) -> Result<(), TransitionError<LogIndex>>;
-
     /// Clones the protocol state as a trait object.
     /// This allows the state to be cloned when it is being used as a `Box<dyn ProtocolSim>`.
     fn clone_box(&self) -> Box<dyn ProtocolSim>;
@@ -156,20 +135,6 @@ pub trait ProtocolSim: std::fmt::Debug + Send + Sync + 'static {
 
 impl Clone for Box<dyn ProtocolSim> {
     fn clone(&self) -> Box<dyn ProtocolSim> {
-        self.clone_box()
-    }
-}
-
-/// ProtocolEvent trait
-///
-/// Defines the interface for protocol-specific events that can be applied to the state.
-pub trait ProtocolEvent: std::fmt::Debug {
-    fn as_any(&self) -> &dyn Any;
-    fn clone_box(&self) -> Box<dyn ProtocolEvent>;
-}
-
-impl Clone for Box<dyn ProtocolEvent> {
-    fn clone(&self) -> Box<dyn ProtocolEvent> {
         self.clone_box()
     }
 }